// apps/web-ui/src/app/services/accounting-api-service/models/entry.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Double-entry accounting ledger entry
 */
export interface Entry {
  credit_amount: string;
  debit_amount: string;
  description: string;
  id: string;
  ledger_accounts_id: string;
  postings_id: string;
  system_notes: string;
}